Base Upload of IdleSoles
You are welcome to expand on the project throughout the semester. I will review any pull requests and enhance the project with your feedback. Note if multiple students work on the same feature, I may not be able to incorporate everyone's code. It may be worthwhile to collaborate on upgrades beforehand.